feat(views/confirmation): implements container hook pattern and states of loading/error
