+++ src/views/ConfirmationView/confirmationView.module.scss
@@ -24,6 +24,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 100%;
 
   &_heading {
     font-size: var(--heading-size-4);
