+++ src/views/ConfirmationView/index.tsx
@@ -1,29 +1,19 @@
 'use client';
 import Link from 'next/link';
 import Head from 'next/head';
-import { useRouter } from 'next/router';
-import { useCallback } from 'react';
-import { useLang } from '@/contexts/langContext';
 import { CheckIcon } from '@/components/icons/CheckIcon';
 import { Button } from '@/components/Button';
 import { PurchaseData } from '@/components/PurchaseData';
 import { userDataMock } from '@/mock/userData';
-import { usePurchaseQuery } from '@/api/hooks/usePurchaseQuery';
 
+import { useConfirmationView } from './useConfirmationView';
+import { Sad } from '@/components/icons/Sad';
+import { ConfirmationSkeleton } from './ConfirmationSkeleton';
 import styles from './confirmationView.module.scss';
 
 export const ConfirmationView = () => {
-  const {
-    lang: { confirmationPage, app },
-  } = useLang();
-  const router = useRouter();
-  const { id } = router.query;
-
-  const goToHome = useCallback(() => {
-    router.push(confirmationPage.goToHome.path);
-  }, [router, confirmationPage]);
-
-  const { data: purchase } = usePurchaseQuery(id as string);
+  const { goToHome, purchase, app, confirmationPage, error, isLoading } =
+    useConfirmationView();
 
   return (
     <main className={styles.container}>
@@ -35,25 +25,45 @@ export const ConfirmationView = () => {
           content={confirmationPage.seo.description}
         />
       </Head>
-      <div className={styles.confirmationHead}>
-        <CheckIcon ariaLabel="check success icon" role="img" />
-        <h1 className={styles.confirmationHead_heading}>
-          {confirmationPage.heading}
-        </h1>
-        <p className={styles.confirmationHead_description}>
-          {confirmationPage.description}
-        </p>
-      </div>
+      {isLoading && <ConfirmationSkeleton />}
 
-      <PurchaseData purchase={purchase} userData={userDataMock} />
+      {!isLoading && !!error && (
+        <div className={styles.confirmationHead}>
+          <Sad />
+          <p className={styles.confirmationHead_heading}>
+            {confirmationPage.errorTitle}
+          </p>
+          <p className={styles.confirmationHead_description}>
+            {confirmationPage.errorDescription}
+          </p>
+        </div>
+      )}
+
+      {!isLoading && purchase && (
+        <>
+          <div className={styles.confirmationHead}>
+            <CheckIcon ariaLabel="check success icon" role="img" />
+            <h1 className={styles.confirmationHead_heading}>
+              {confirmationPage.heading}
+            </h1>
+            <p className={styles.confirmationHead_description}>
+              {confirmationPage.description}
+            </p>
+          </div>
+
+          <PurchaseData purchase={purchase} userData={userDataMock} />
+        </>
+      )}
 
       <footer className={styles.actions}>
-        <Link
-          className={styles.actions_manageSubscription}
-          href={confirmationPage.manageSubscription.path}
-        >
-          {confirmationPage.manageSubscription.label}
-        </Link>
+        {!isLoading && purchase && (
+          <Link
+            className={styles.actions_manageSubscription}
+            href={confirmationPage.manageSubscription.path}
+          >
+            {confirmationPage.manageSubscription.label}
+          </Link>
+        )}
         <Button className={styles.actions_goToHome} onClick={goToHome}>
           {confirmationPage.goToHome.label}
         </Button>
